mgr/cephadm: move _run_cephadm to serve.py
